# bpo-29312: use METH_FASTCALL for dict.update