@jdemeyer 2019-07-08
Here are some benchmarks:
1. `./python -m perf timeit --duplicate 200 -s 'L = [("x",None)]' 'dict(L)'`
**before**: Mean +- std dev: 221 ns +- 1 ns
**after**: Mean +- std dev: 186 ns +- 1 ns
2. `./python -m perf timeit --duplicate 200 -s 'L = [("x",None)]' '{}.update(L)'`
**before**: Mean +- std dev: 189 ns +- 2 ns
**after**: Mean +- std dev: 161 ns +- 1 ns
3. `./python -m perf timeit --duplicate 200 -s 'D = {"x":None}' 'dict(D)'`
**before**: Mean +- std dev: 196 ns +- 1 ns
**after**: Mean +- std dev: 126 ns +- 1 ns
4. `./python -m perf timeit --duplicate 200 -s 'D = {"x":None}' '{}.update(D)'`
**before**: Mean +- std dev: 174 ns +- 1 ns
**after**: Mean +- std dev: 116 ns +- 0 ns
5. `./python -m perf timeit --duplicate 200 'dict(x=None)'`
**before**: Mean +- std dev: 200 ns +- 1 ns
**after**: Mean +- std dev: 119 ns +- 1 ns
6. `./python -m perf timeit --duplicate 200 '{}.update(x=None)'`
**before**: Mean +- std dev: 201 ns +- 1 ns
**after**: Mean +- std dev: 120 ns +- 1 ns
7. `./python -m perf timeit --duplicate 200 -s 'D = {"x":None}' 'dict(**D)'`
**before**: Mean +- std dev: 149 ns +- 1 ns
**after**: Mean +- std dev: 171 ns +- 1 ns
8. `./python -m perf timeit --duplicate 200 -s 'D = {"x":None}' '{}.update(**D)'`
**before**: Mean +- std dev: 139 ns +- 1 ns
**after**: Mean +- std dev: 171 ns +- 1 ns
9. `./python -m perf timeit --duplicate 200 -s 'D = {f"x{i}":None for i in range(10000)}' 'dict(**D)'`
**before**: Mean +- std dev: 321 us +- 1 us
**after**: Mean +- std dev: 393 us +- 3 us
10. `./python -m perf timeit --duplicate 200 -s 'D = {f"x{i}":None for i in range(10000)}' '{}.update(**D)'`
**before**: Mean +- std dev: 321 us +- 3 us
**after**: Mean +- std dev: 394 us +- 4 us

So the calls `dict(**D)` and `dict.update(**D)` got slower but the rest got faster.