@jdemeyer 2019-07-09
I changed the branch to match the title, reverting the changes to `__init__`.

Note that this PR has more code duplication: the code paths for `dict.update` and `dict.__init__` used to be the same but that's no longer the case with this PR. When `dict.__init__` is also changed (as in my original PR), the code paths will be the same again.