@jdemeyer 2019-08-16
Rebased to latest master, using #14682.