@vstinner 2019-08-19
@methane: I let you decide what do to do with this change.